feat: Pull sheet checking into a separate file
